+++ 2D/Belyi_Torus.cpp
@@ -16,16 +16,16 @@
 
 using namespace vb; using namespace std;
 
-vector<unsigned> ntri { 0, 1, 5, 46, 669 };
+vector<int> ntri { 0, 1, 5, 46, 669 };
 
 int main (int argc, char ** argv) {
 	H.init ("Toroidal enumeration", argc, argv, "s=1,m=228,r=1,o,d=0,D=0,g=1,f,n=2,q");
 	int s=H['s'], g=H['g'], a=6*(s+2*g-2), r=H['r'], d=H['d'], D=H['D'];
 	assert (a>0); if (g!=1) assert(!H['o']); if (r>0) prng.seed(r);
 
-	Cycles phi_c; for (unsigned long i=0; i<a/3; ++i) phi_c.emplace_back (std::vector<unsigned long> {3*i,3*i+1,3*i+2}); Permutation phi(phi_c);
+	Cycles phi_c; for (int i=0; i<a/3; ++i) phi_c.emplace_back (std::vector<int> {3*i,3*i+1,3*i+2}); Permutation phi(phi_c);
 
-	vector<Hypermap> v; unsigned target = 0; if ((d==0) && (g==1) && (!H['f']) && (s<ntri.size())) target = ntri[s];
+	vector<Hypermap> v; int target = 0; if ((d==0) && (g==1) && (!H['f']) && (s<ntri.size())) target = ntri[s];
 
 	Coloring img (cpx(-1,-1),cpx(1,1),500,[](cpx){return BLACK;});
